renamed NestedAlertContent files
