--- packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/NestedAlertContentNoProduct.js
@@ -1,13 +1,0 @@
-import React from "react";
-import { NestedAlertContentEdgeCase } from "./NestedAlertContentEdgeCase";
-
-const NestedAlertContentNoProduct = ({content}) => {
-
-  return (
-    <>
-      <NestedAlertContentEdgeCase content={content} isNoProduct={true} />
-    </>
-  );
-};
-
-export { NestedAlertContentNoProduct };
