--- packages/gatsby-theme-aio/src/components/GetCredential/RequestAccess/NestedAlertContentNotMember.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { NestedAlertContentEdgeCase } from './NestedAlertContentEdgeCase';
-
-const NestedAlertContentNotMember = ({content}) => {
-
-  return (
-    <>
-      <NestedAlertContentEdgeCase content={content} />
-    </>
-  );
-};
-
-export { NestedAlertContentNotMember };
