+++ packages/gatsby-theme-aio/src/components/GetCredential/index.js
@@ -48,11 +48,11 @@ import { RequestAccessSide } from "./RequestAccess/RequestAccessSide";
 import Context from '../Context';
 import GetCredentialContext from './GetCredentialContext';
 import { ReturnCredentialDetails } from './Return/ReturnCredentialDetails';
-import { NestedAlertContentEdgeCase } from './RequestAccess/NestedAlertContentEdgeCase';
-import { NestedAlertContentNoProduct } from './RequestAccess/NestedAlertContentNoProduct';
-import { NestedAlertContentType1User } from './RequestAccess/NestedAlertContentType1User';
-import { NestedAlertContentNotMember } from './RequestAccess/NestedAlertContentNotMember';
-import { NestedAlertContentNotSignUp } from './RequestAccess/NestedAlertContentNotSignUp';
+import { OrganizationAccessDetailsEdgeCase } from './RequestAccess/OrganizationAccessDetailsEdgeCase';
+import { OrganizationAccessDetailsNoProduct } from './RequestAccess/OrganizationAccessDetailsNoProduct';
+import { OrganizationAccessDetailsType1User } from './RequestAccess/OrganizationAccessDetailsType1User';
+import { OrganizationAccessDetailsNotMember } from './RequestAccess/OrganizationAccessDetailsNotMember';
+import { OrganizationAccessDetailsNotSignUp } from './RequestAccess/OrganizationAccessDetailsNotSignUp';
 
 const LocalStorageKey = 'OrgInfo';
 
@@ -386,11 +386,11 @@ GetCredential.Return.CredentialDetails.AllowedOrigins = ReturnAllowedOrigins;
 GetCredential.Return.CredentialDetails.APIKey = ReturnAPIKey;
 GetCredential.RequestAccess = RequestAccess;
 GetCredential.RequestAccess.RestrictedAccess = RestrictedAccess;
-GetCredential.RequestAccess.EdgeCase = NestedAlertContentEdgeCase;
-GetCredential.RequestAccess.EdgeCase.NoProduct = NestedAlertContentNoProduct;
-GetCredential.RequestAccess.EdgeCase.Type1User = NestedAlertContentType1User;
-GetCredential.RequestAccess.EdgeCase.NotMember = NestedAlertContentNotMember;
-GetCredential.RequestAccess.EdgeCase.NotSignUp = NestedAlertContentNotSignUp;
+GetCredential.RequestAccess.EdgeCase = OrganizationAccessDetailsEdgeCase;
+GetCredential.RequestAccess.EdgeCase.NoProduct = OrganizationAccessDetailsNoProduct;
+GetCredential.RequestAccess.EdgeCase.Type1User = OrganizationAccessDetailsType1User;
+GetCredential.RequestAccess.EdgeCase.NotMember = OrganizationAccessDetailsNotMember;
+GetCredential.RequestAccess.EdgeCase.NotSignUp = OrganizationAccessDetailsNotSignUp;
 GetCredential.RequestAccess.RestrictedAccess.Products = Products;
 GetCredential.RequestAccess.RestrictedAccess.Products.Product = Product;
 GetCredential.RequestAccess.RequestAccessSide = RequestAccessSide;
